+++ bench/Main.hs
@@ -1,12 +1,10 @@
 import Control.Varying
+import Control.Monad
 import Criterion.Main
-import Debug.Trace
 
 main :: IO ()
 main = do
-    let v :: Var Int Int
-        v = var (+1)
-        run v a = fst <$> runVarT v a
+    let run v a = fst <$> runVarT v a
     defaultMain $ [ bgroup "runVarT" [ bench "1" $ nf (run $ chain 1) 0
                                      , bench "2" $ nf (run $ chain 2) 0
                                      , bench "4" $ nf (run $ chain 4) 0
@@ -16,9 +14,19 @@ main = do
                                      , bench "64" $ nf (run $ chain 64) 0
                                      , bench "128" $ nf (run $ chain 128) 0
                                      ]
+                  , bgroup "SplineT"
+                      [ bench "runSplineT" $
+                          nf (run $ outputStream spline 0) 0
+                      ]
                   ]
     return ()
 
 chain :: Int -> Var Int Int
 chain n = seq x x
   where x = foldl (~>) (var (+1)) $ take (n - 1) $ cycle [var (+1)]
+
+spline :: Spline Float Float ()
+spline = do
+  void $ tween easeInExpo 0 100 1
+  void $ tween easeOutExpo 100 0 1
+  spline