+++ changelog.md
@@ -11,6 +11,9 @@ change log
 0.3.1.0 - added stepMany, eitherE
 
 0.4.0.0 - Var and Spline are now parameterized with Identity, removed mix, changed
-          the behavior of race, added untilEvent variants, added tests.
+          the behavior of race, added untilEvent variants, added tests
 
-0.5.0.0 - changed stepMany to remove Monoid requirement
+0.5.0.0 - changed stepMany to remove Monoid requirement, added raceMany, added 
+          anyE, more tests and SplineT obeys Applicative and Monad laws
+
+0.5.0.1 - removed time as dependency